requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-QRcode==3.1.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pillow==10.2.0
pypng==0.20220715.0
qrcode==7.4.2
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
